Removed Perturbation between beta
Due to this change, I had to change the range of beta to 0.1 -> 1e6. The question is, can we not increase the range, only the steps of Y? And is there a guarantee that Y locations don't change if I increase Beta further? Right now the current running time is ~ 8 mins